DNAZooData

Fetches files from the DNAZoo data portal and caches them using the BiocFileCache system.

Arguments

species

Any species processed by the DNA Zoo (check https://www.dnazoo.org/assemblies) for a browser-based explorer.

Value

DNAZooData() returns a HicFile object, which can then be imported in memory using HiCExperiment::import(). Metadata also points to a URL to directly fetch the genome assembly corrected by the DNA Zoo consortium.

Examples

###################################
## Importing DNAZoo `.hic` files ##
###################################

head(DNAZooData())
#>                   species                              readme
#> 1        Acinonyx_jubatus        Acinonyx_jubatus/README.json
#> 2      Acropora_millepora      Acropora_millepora/README.json
#> 3     Addax_nasomaculatus     Addax_nasomaculatus/README.json
#> 4           Aedes_aegypti           Aedes_aegypti/README.json
#> 5   Aedes_aegypti__AaegL4   Aedes_aegypti__AaegL4/README.json
#> 6 Aedes_aegypti__AaegL5.0 Aedes_aegypti__AaegL5.0/README.json
#>                                                           readme_link
#> 1        https://dnazoo.s3.wasabisys.com/Acinonyx_jubatus/README.json
#> 2      https://dnazoo.s3.wasabisys.com/Acropora_millepora/README.json
#> 3     https://dnazoo.s3.wasabisys.com/Addax_nasomaculatus/README.json
#> 4           https://dnazoo.s3.wasabisys.com/Aedes_aegypti/README.json
#> 5   https://dnazoo.s3.wasabisys.com/Aedes_aegypti__AaegL4/README.json
#> 6 https://dnazoo.s3.wasabisys.com/Aedes_aegypti__AaegL5.0/README.json
#>   original_assembly     new_assembly
#> 1           aciJub1      aciJub1_HiC
#> 2       amil_sf_1.1  amil_sf_1.1_HiC
#> 3      ASM1959352v1 ASM1959352v1_HiC
#> 4        AGWG.draft         AaegL5.0
#> 5            AaegL3           AaegL4
#> 6        AGWG.draft         AaegL5.0
#>                                                               new_assembly_link
#> 1         https://dnazoo.s3.wasabisys.com/Acinonyx_jubatus/aciJub1_HiC.fasta.gz
#> 2   https://dnazoo.s3.wasabisys.com/Acropora_millepora/amil_sf_1.1_HiC.fasta.gz
#> 3 https://dnazoo.s3.wasabisys.com/Addax_nasomaculatus/ASM1959352v1_HiC.fasta.gz
#> 4               https://dnazoo.s3.wasabisys.com/Aedes_aegypti/AaegL5.0.fasta.gz
#> 5         https://dnazoo.s3.wasabisys.com/Aedes_aegypti__AaegL4/AaegL4.fasta.gz
#> 6     https://dnazoo.s3.wasabisys.com/Aedes_aegypti__AaegL5.0/AaegL5.0.fasta.gz
#>   new_assembly_link_status
#> 1                      200
#> 2                      200
#> 3                      200
#> 4                      404
#> 5                      200
#> 6                      200
#>                                                                   hic_link
#> 1    https://dnazoo.s3.wasabisys.com/Acinonyx_jubatus/aciJub1.rawchrom.hic
#> 2   https://dnazoo.s3.wasabisys.com/Acropora_millepora/amil_sf_1.1_HiC.hic
#> 3 https://dnazoo.s3.wasabisys.com/Addax_nasomaculatus/ASM1959352v1_HiC.hic
#> 4                                                                     <NA>
#> 5         https://dnazoo.s3.wasabisys.com/Aedes_aegypti__AaegL4/AaegL4.hic
#> 6     https://dnazoo.s3.wasabisys.com/Aedes_aegypti__AaegL5.0/AaegL5.0.hic
hf <- DNAZooData(species = 'Anolis_carolinensis')
hf
#> HicFile object
#> .hic file: /github/home/.cache/R/DNAZooData/665fe1f8950_AnoCar2.0_HiC.hic 
#> resolution: 5000 
#> pairs file: 
#> metadata(6): organism draftAssembly ... credits assemblyURL